Update L_Arduino.lua
There was a bug in updating HVAC device on vera. it kept reseting the setpoint degrees to 0.0 each time a command was sent from Vera to the arduino HVAC device. and alos not updating the current setpoint when sending a message with new setpoint from HVAC device to Vera.
Adding lines 219-221 solved this.

To the Admins, U can add me as an editor for future bugs I will find and fix.